order and manage user module

+++ minimartdemo/FE/src/features/User/components/Address/index.jsx
@@ -111,10 +111,12 @@ function Address() {
 						<p onClick={() => onClickEditBtn(item)}>Sửa</p>
 						<p
 							onClick={() => {
-								window.prompt(
-									"Bạn có chắc muốn xóa địa chỉ này"
-								);
-								dispatch(deleteUserAddress(item._id));
+								if (
+									window.confirm(
+										"Bạn có chắc muốn xóa địa chỉ này"
+									)
+								)
+									dispatch(deleteUserAddress(item._id));
 							}}>
 							Xóa
 						</p>
